Updated all input form pages to check for unsaved changes and display message if pages are unsaved and a page exit is tried
Note: for "Edit Work Experiences" page, there needed to be 2 buttons just to ensure that the "unsaved pages" blocker doesn't persist when the page is saved at the bottom. The "Save Responsibilities" button removes the blocker for the Responsibilities parts affected by that function while the real "Save" button saves everything to the backend and removes the blocker for the inputs in the rest of the page. User is instructed to save with both buttons so that the info is still saved on the backend despite the work around. Had to do it this way because I did not want to mess with how the functions interact with the Gemini API as is and cause massive bugs